refactor(app): extract manage user page controller (#346)
Extract ManageUser.vue's setup() page-level composition (bulk-selection
wiring, mount/URL-restore lifecycle, filter/sort watchers, and the legacy
action aliases) into views/admin/composables/useManageUserPage.ts, mirroring
the useEntitiesTable.ts thin-shell pattern (`setup() { return
useManageUserPage(); }`). useUserData, useUserMutations,
useBulkUserActions, useUserModals, and useUserTablePresentation remain the
sole owners of their existing responsibilities and are only wired together
by the new composable.

Adds ManageUser.spec.ts coverage for behaviours that were previously only
implicit: the filter-preset seed running exactly once (idempotent across
remounts), the 20-user select-all-on-page cap and its warning toast, the
last-Administrator bulk-delete guard, and the full legacy alias surface
(confirmDeleteUser, updateUserData, confirmBulkApprove,
confirmBulkRoleAssignment, confirmBulkDelete, changeUserPassword,
makeToast).

ManageUser.vue: 611 -> 372 lines. New useManageUserPage.ts: 270 lines.
ManageUser.spec.ts stays under 600 (597) via doc-comment trims alongside
the new test coverage.
